+++ packages/shared/src/types/billing.ts
@@ -1034,8 +1034,7 @@ interface CheckoutFlowMethods {
 }
 
 /**
- * Public API type for checkout flow. Properties are null when status is 'needs_initialization'
- * and present when status is 'needs_confirmation' or 'completed'.
+ * @interface
  */
 export type CheckoutFlowResource = CheckoutPropertiesPerStatus & CheckoutFlowMethods;